Move user related role api to AuthContainer
