+++ packages/skygear-core/lib/auth.js
@@ -13,8 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+const _ = require('lodash');
+
 import {EventHandle} from './util';
 import {ErrorCodes} from './error';
+import Role from './role';
 
 export const USER_CHANGED = 'userChanged';
 
@@ -132,6 +135,79 @@ export class AuthContainer {
       password: newPassword
     })
     .then(this._authResolve.bind(this));
+  }
+
+  setAdminRole(roles) {
+    let roleNames = _.map(roles, function (perRole) {
+      return perRole.name;
+    });
+
+    return this.container.makeRequest('role:admin', {
+      roles: roleNames
+    }).then((body) => body.result);
+  }
+
+  setDefaultRole(roles) {
+    let roleNames = _.map(roles, function (perRole) {
+      return perRole.name;
+    });
+
+    return this.container.makeRequest('role:default', {
+      roles: roleNames
+    }).then((body) => body.result);
+  }
+
+  fetchUserRole(users) {
+    let userIds = _.map(users, function (perUser) {
+      // accept either user record or user id
+      return perUser._id || perUser;
+    });
+
+    return this.container.makeRequest('role:get', {
+      users: userIds
+    })
+    .then((body) =>
+      Object.keys(body.result)
+      .map((key) => [key, body.result[key]])
+      .reduce((acc, pairs) => ({
+        ...acc || {},
+        [pairs[0]]: pairs[1].map((name) => new Role(name))
+      }), null)
+    );
+  }
+
+  assignUserRole(users, roles) {
+    let userIds = _.map(users, function (perUser) {
+      // accept either user record or user id
+      return perUser._id || perUser;
+    });
+
+    let roleNames = _.map(roles, function (perRole) {
+      // accept either role object or role name
+      return perRole.name || perRole;
+    });
+
+    return this.container.makeRequest('role:assign', {
+      users: userIds,
+      roles: roleNames
+    }).then((body) => body.result);
+  }
+
+  revokeUserRole(users, roles) {
+    let userIds = _.map(users, function (perUser) {
+      // accept either user record or user id
+      return perUser._id || perUser;
+    });
+
+    let roleNames = _.map(roles, function (perRole) {
+      // accept either role object or role name
+      return perRole.name || perRole;
+    });
+
+    return this.container.makeRequest('role:revoke', {
+      users: userIds,
+      roles: roleNames
+    }).then((body) => body.result);
   }
 
   _getAccessToken() {
